Move/rename and create superclass for unit tests
