+++ test/fixtures/basic/configmap-data.yml
@@ -1,7 +1,7 @@
 apiVersion: v1
 kind: ConfigMap
 metadata:
-  name: configmap-data
+  name: basic-configmap-data
 data:
   datapoint1: value1
   datapoint2: value2
