+++ test/fixtures/basic/unmanaged-pod.yml.erb
@@ -17,5 +17,5 @@ spec:
       - name: CONFIG
         valueFrom:
           configMapKeyRef:
-            name: configmap-data
+            name: basic-configmap-data
             key: datapoint2
