+++ test/fixtures/basic/web.yml.erb
@@ -54,5 +54,5 @@ spec:
         - name: CONFIG
           valueFrom:
             configMapKeyRef:
-              name: configmap-data
+              name: basic-configmap-data
               key: datapoint1
